+++ docs/changelog.rst
@@ -2,8 +2,8 @@
 Changelog
 =========
 
-Changes in 0.6
-==============
+Changes in 0.6.x
+================
 
 - Added FutureWarning to inherited classes not declaring 'allow_inheritance' as the default will change in 0.7
 - Added support for covered indexes when inheritance is off